Move unused lambda schedule to attic
Was optimized for earlier iterations of RBFE and RABFE protocol
