+++ timemachine/fe/lambda_schedule.py
@@ -1,34 +1,6 @@
 import numpy as np
 
 
-def construct_lambda_schedule(num_windows):
-    """Generate a length-num_windows list of lambda values from 0.0 up to 1.0
-
-    Notes
-    -----
-    manually optimized by YTZ
-    """
-
-    A = int(0.35 * num_windows)
-    B = int(0.30 * num_windows)
-    C = num_windows - A - B
-
-    # Empirically, we see the largest variance in std <du/dl> near the endpoints in the nonbonded
-    # terms. Bonded terms are roughly linear. So we add more lambda windows at the endpoint to
-    # help improve convergence.
-    lambda_schedule = np.concatenate(
-        [
-            np.linspace(0.0, 0.25, A, endpoint=False),
-            np.linspace(0.25, 0.75, B, endpoint=False),
-            np.linspace(0.75, 1.0, C, endpoint=True),
-        ]
-    )
-
-    assert len(lambda_schedule) == num_windows
-
-    return lambda_schedule
-
-
 def validate_lambda_schedule(lambda_schedule, num_windows):
     """Must go monotonically from 0 to 1 in num_windows steps"""
     assert lambda_schedule[0] == 0.0
